requirement.txt
streamlit
azure-storage-blob
azure-identity